ufm: Fix writing incorrect offset if offset exceeds 0x100
Root cause:
It declares the parameter of addr to uint8_t in set_provision_data_in_flash API and the range of its value is from 0 to 255.
Ex: if users want to write the data at 0x12c offset, the data will be accidently wrote at 0x2c offset.

Solution:
Change to declare this parameter to uint32_t and removes redundant code.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>
Change-Id: I8a92cea9b94bb43e14a92c30ed7384b8c05f8c4c