spdm: generate incorrect MEASUREMENTS message
The signature of ecdsa r and s could be leading with 0x00, and
mbedtls_mpi_size() will not including this byte. So user has to padding
to it's own binary. Refactor to a fixed length instead.

Refactor the signature generation and verification for future use.

AST1060 <-> AST2600 measurement takes 3 second in both directions.

Signed-off-by: Troy Lee <troy_lee@aspeedtech.com>
Change-Id: I5268876265276f8f63b3897208a6b6b7059717a7