afm: Adjust default AFM capsule size to 4KB
The Intel Prot tool (intel-server-prot-spdm) uses 4KB as the default
AFM capsule size. To stay compatible and avoid issues when the tool is
updated, change the AFM default size from 8KB to 4KB.

This patch also updates the documentation to explain how to use the
Intel Prot tool to generate signed images and various capsules for
different test scenarios.

Signed-off-by: Kane-Chen-AS <kane_chen@aspeedtech.com>
Change-Id: Iba34e9afd44bba0758d2b05272f05000c348a172