mcu-runtime: spi: Support single flash ABR mode
Add support for single flash ABR mode in the SPI loader.
When ABR is enabled and single flash mode is detected,
the image source offset is adjusted by adding half of
the physical flash size.

This ensures the correct image is loaded from the backup
region in single flash configurations.

Signed-off-by: Chin-Ting Kuo <chin-ting_kuo@aspeedtech.com>
Change-Id: Ib535c930a1f4c86d3e24c09b937025a2b955fba5