pci: Use dts to configure pcie-ep devices
There are multiple pcie-ep behind aspeed bridge including vga, bmc-dev,
ehci, xhci. We check the properties in dts to tell if they should be
enabled.

They are default enabled in the dts now. Delete any of them you do not
need. The value of the property define the interrupt mode, msi/intx.

Ex.
&pcie0 {
	/delete-property/ xhci;
	xhci = "MSI";
};

Signed-off-by: Jammy Huang <jammy_huang@aspeedtech.com>
Change-Id: Id0bbc15691a18ca73b7b0285c018696d1155db7d